refactor: [MEW-1937] extract safeParseUnits to shared utils/unit
Move the comma-normalizing parseUnits wrapper from ModuleSend into
src/utils/unit.ts so it can be reused by ModuleSwap and ModuleTrade.
